+++ server/src/main/java/com/onetool/server/api/member/controller/MemberEmailController.java
@@ -2,8 +2,8 @@ package com.onetool.server.api.member.controller;
 
 import com.onetool.server.api.member.business.MemberBusiness;
 import com.onetool.server.api.member.business.MemberEmailBusiness;
-import com.onetool.server.api.member.dto.MemberFindEmailRequest;
-import com.onetool.server.api.member.dto.MemberFindPwdRequest;
+import com.onetool.server.api.member.dto.request.MemberFindEmailRequest;
+import com.onetool.server.api.member.dto.request.MemberFindPwdRequest;
 import com.onetool.server.global.exception.ApiResponse;
 import jakarta.validation.Valid;
 import lombok.AllArgsConstructor;
