+++ server/src/main/java/com/onetool/server/api/member/dto/request/MemberFindEmailRequest.java
@@ -1,4 +1,4 @@
-package com.onetool.server.api.member.dto;
+package com.onetool.server.api.member.dto.request;
 
 public record MemberFindEmailRequest(
         String name,
