+++ server/src/main/java/com/onetool/server/api/member/dto/response/MemberInfoResponse.java
@@ -1,4 +1,4 @@
-package com.onetool.server.api.member.dto;
+package com.onetool.server.api.member.dto.response;
 
 import com.onetool.server.api.member.domain.Member;
 import com.onetool.server.api.order.Orders;
