+++ server/src/main/java/com/onetool/server/api/member/dto/response/MemberLoginResponse.java
@@ -1,4 +1,4 @@
-package com.onetool.server.api.member.dto;
+package com.onetool.server.api.member.dto.response;
 
 import lombok.Builder;
