+++ server/src/main/java/com/onetool/server/global/auth/login/handler/OAuth2LoginSuccessHandler.java
@@ -3,7 +3,7 @@ package com.onetool.server.global.auth.login.handler;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.onetool.server.global.auth.MemberAuthContext;
 import com.onetool.server.global.auth.jwt.JwtUtil;
-import com.onetool.server.api.member.dto.MemberLoginResponse;
+import com.onetool.server.api.member.dto.response.MemberLoginResponse;
 import com.onetool.server.api.member.repository.MemberRepository;
 import com.onetool.server.api.member.enums.UserRole;
 import com.onetool.server.global.auth.login.PrincipalDetails;
