+++ server/src/test/java/com/onetool/server/api/member/controller/MemberControllerTest.java
@@ -3,10 +3,10 @@ package com.onetool.server.api.member.controller;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.onetool.server.api.helper.MockBeanInjection;
 import com.onetool.server.api.member.domain.Member;
-import com.onetool.server.api.member.dto.MemberCreateRequest;
-import com.onetool.server.api.member.dto.MemberCreateResponse;
-import com.onetool.server.api.member.dto.MemberInfoResponse;
-import com.onetool.server.api.member.dto.MemberUpdateRequest;
+import com.onetool.server.api.member.dto.request.MemberCreateRequest;
+import com.onetool.server.api.member.dto.response.MemberCreateResponse;
+import com.onetool.server.api.member.dto.response.MemberInfoResponse;
+import com.onetool.server.api.member.dto.request.MemberUpdateRequest;
 import com.onetool.server.api.member.fixture.MemberFixture;
 import com.onetool.server.api.member.fixture.WithMockPrincipalDetails;
 import com.onetool.server.global.auth.MemberAuthContext;
